@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
@@ -32,14 +32,14 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.3
   rxdart: ^0.23.0
-  http:
-  provider:
-  dio:
-  image_picker:
-  mime_type:
-  shared_preferences:
-  flutter_swiper: ^1.1.6
-  animations: ^1.1.2
+  http: ^0.13.3
+  dio: ^4.0.0
+  card_swiper: ^1.0.4
+  animations: ^2.0.1
+  firebase_messaging: ^10.0.4
+  firebase_core: ^1.4.0
+  flutter_secure_storage: ^4.2.1
+  provider: ^5.0.0
 
 dev_dependencies:
   flutter_test:
