@@ -1,21 +1,22 @@
 import 'dart:convert';
 
-SearchRegion searchRegionFromJson(String str) => SearchRegion.fromJson(json.decode(str));
+SearchRegion searchRegionFromJson(String str) =>
+    SearchRegion.fromJson(json.decode(str));
 
 String searchRegionToJson(SearchRegion data) => json.encode(data.toJson());
 
 class SearchRegion {
-    SearchRegion({
-        this.searchWord,
-    });
+  SearchRegion({
+    this.searchWord,
+  });
 
-    String searchWord;
+  String? searchWord;
 
-    factory SearchRegion.fromJson(Map<String, dynamic> json) => SearchRegion(
+  factory SearchRegion.fromJson(Map<String, dynamic> json) => SearchRegion(
         searchWord: json["search_word"],
-    );
+      );
 
-    Map<String, dynamic> toJson() => {
+  Map<String, dynamic> toJson() => {
         "search_word": searchWord,
-    };
+      };
 }
