@@ -7,7 +7,7 @@ import 'package:traveling/src/pages/search.dart';
 import 'package:traveling/src/pages/travels.dart';
 
 class MainScreen extends StatefulWidget {
-  MainScreen({Key key}) : super(key: key);
+  MainScreen({Key? key}) : super(key: key);
 
   @override
   _MainScreenState createState() => _MainScreenState();
@@ -22,6 +22,7 @@ class _MainScreenState extends State<MainScreen> {
     TravelsPage(),
     ProfilePage()
   ];
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
