@@ -1,19 +1,18 @@
+import 'package:flutter/cupertino.dart';
 import 'package:http/http.dart' as http;
+import 'package:provider/provider.dart';
 import 'package:traveling/src/models/acceptOffer_model.dart';
 import 'package:traveling/src/models/confirmOffer_model.dart';
 import 'package:traveling/src/models/flightOffer_model.dart';
 import 'package:traveling/src/models/searchTicket_model.dart';
 import 'package:traveling/src/models/ticketList_model.dart';
-import 'package:traveling/src/preferences/user_preferences.dart';
+import 'package:traveling/src/providers/UserProvider.dart';
 import 'package:traveling/src/providers/global.dart';
 
 class TicketProvider {
-  final String _url = baseUrl;
-
   Future<List<TicketList>> searchTickets(SearchTicket searchTicket) async {
-    final url = '$_url/obtener-vuelos-disponibles/';
-
-    final response = await http.post(url,
+    final response = await http.post(
+        Uri.parse('$baseUrl/obtener-vuelos-disponibles/'),
         headers: {
           'Content-Type': 'application/json',
           'Accept': 'application/json'
@@ -27,10 +26,9 @@ class TicketProvider {
     return [];
   }
 
-  Future<ConfirmOffer> confirmTicket(FlightOffer flightOffer) async {
-    final url = '$_url/confirmar-precio-vuelo/';
-
-    final response = await http.post(url,
+  Future<ConfirmOffer?> confirmTicket(FlightOffer flightOffer) async {
+    final response = await http.post(
+        Uri.parse('$baseUrl/confirmar-precio-vuelo/'),
         headers: {
           'Content-Type': 'application/json',
           'Accept': 'application/json'
@@ -44,12 +42,12 @@ class TicketProvider {
     return null;
   }
 
-  Future<bool> acceptTicket(AcceptOffer acceptOffer) async {
-    final url = '$_url/crear-reserva/';
-    final UserPreferences userPreferences = new UserPreferences();
-
-    final response = await http.post(url,
-        headers: {'Authorization': 'Bearer ' + userPreferences.accessToken},
+  Future<bool> acceptTicket(
+      AcceptOffer acceptOffer, BuildContext context) async {
+    final userProvider = Provider.of<UserProvider>(context, listen: false);
+    final accessToken = await userProvider.readAccessToken();
+    final response = await http.post(Uri.parse('$baseUrl/crear-reserva/'),
+        headers: {'Authorization': 'Bearer ' + accessToken},
         body: acceptOfferToJson(acceptOffer));
 
     if (response.statusCode == 200) {
