@@ -1,17 +1,15 @@
 import 'dart:convert';
+import 'package:flutter/material.dart';
+import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:http/http.dart' as http;
 import 'package:traveling/src/models/user_model.dart';
-import 'package:traveling/src/preferences/user_preferences.dart';
 import 'package:traveling/src/providers/global.dart';
 
-class UserProvider {
-  final String _url = baseUrl;
-  final _preferences = new UserPreferences();
+class UserProvider extends ChangeNotifier {
+  final storage = FlutterSecureStorage();
 
   Future<bool> login(User user) async {
-    final url = '$_url/iniciar-sesion/';
-
-    final response = await http.post(url,
+    final response = await http.post(Uri.parse('$baseUrl/iniciar-sesion/'),
         headers: {
           'Content-Type': 'application/json',
           'Accept': 'application/json'
@@ -20,18 +18,15 @@ class UserProvider {
 
     if (response.statusCode == 200) {
       final data = json.decode(response.body);
-      _preferences.accessToken = data["access_token"];
-      _preferences.refreshToken = data["refresh_token"];
-      _preferences.lastPage = 'main';
+      await storage.write(key: 'access_token', value: data["access_token"]);
+      await storage.write(key: 'refresh_token', value: data["refresh_token"]);
       return true;
     }
     return false;
   }
 
   Future<bool> updatePassword(User user) async {
-    final url = '$_url/cambiar-contrasena/';
-
-    final response = await http.post(url,
+    final response = await http.post(Uri.parse('$baseUrl/cambiar-contrasena/'),
         headers: {
           'Content-Type': 'application/json',
           'Accept': 'application/json'
@@ -45,14 +40,31 @@ class UserProvider {
   }
 
   Future<bool> logout(User user) async {
-    final url = '$_url/cerrar-sesion/';
-
-    final response = await http.post(url,
+    final response = await http.post(Uri.parse('$baseUrl/cerrar-sesion/'),
         headers: {'Content-Type': 'application/json'}, body: userToJson(user));
 
     if (response.statusCode == 200) {
+      await storage.deleteAll();
       return true;
     }
     return false;
+  }
+
+  Future clearTokens() async {
+    return await storage.deleteAll();
+  }
+
+  Future<String> readAccessToken() async {
+    return await storage.read(key: 'access_token') ?? '';
+  }
+
+  Future<String> readRefreshToken() async {
+    return await storage.read(key: 'refresh_token') ?? '';
+  }
+
+  Future<List<String>> readToken() async {
+    String accessToken = await storage.read(key: 'access_token') ?? '';
+    String refreshToken = await storage.read(key: 'refresh_token') ?? '';
+    return [accessToken, refreshToken];
   }
 }
